OWLS-99380: Fix the pod roll issue in auxiliary image domain upgrade scenario from 3.4 to 4.0.
